Move signature generation to its own object
